requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
ffmpeg-python==0.2.0
python-multipart==0.0.9
aiofiles==24.1.0
boto3>=1.34.0
websockets>=12.0
slowapi>=0.1.9
python-dotenv>=1.0.0
prometheus-client>=0.20.0
requests>=2.31.0
